+++ CHANGELOG.md
@@ -17,6 +17,7 @@ Given a version number MAJOR.MINOR.PATCH, increment:
 - Normalized event signature verification after failed raw content attempt
 - DictKey.bank_name parameter
 - Workspace.allowed_tax_ids to control who can send Deposits to the Workspace
+- Workspace.update() to allow parameter updates
 
 ## [2.9.0] - 2021-03-22
 ### Added